AS7-3996 deployment crashes console with --header={rollout id=XXX} parameter
