+++ cli/src/main/java/org/jboss/as/cli/ArgumentValueConverter.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source.
- * Copyright 2011, Red Hat, Inc., and individual contributors
+ * Copyright 2012, Red Hat, Inc., and individual contributors
  * as indicated by the @author tags. See the copyright.txt file in the
  * distribution for a full listing of individual contributors.
  *
@@ -22,13 +22,8 @@
 package org.jboss.as.cli;
 
 
-import java.util.Collection;
 
-import org.jboss.as.cli.operation.ParsedOperationRequestHeader;
-import org.jboss.as.cli.operation.impl.DefaultCallbackHandler;
-import org.jboss.as.cli.parsing.DefaultParsingState;
-import org.jboss.as.cli.parsing.ParserUtil;
-import org.jboss.as.cli.parsing.operation.HeaderListState;
+import org.jboss.as.cli.impl.HeadersArgumentValueConverter;
 import org.jboss.dmr.ModelNode;
 
 
@@ -133,30 +128,6 @@ public interface ArgumentValueConverter {
                 o.get(propName).set(prop.substring(equals + 1));
             }
             return o;
-        }
-    };
-
-    ArgumentValueConverter HEADERS = new DMRWithFallbackConverter() {
-        private final DefaultCallbackHandler callback = new DefaultCallbackHandler();
-        private final DefaultParsingState initialState = new DefaultParsingState("INITIAL_STATE");
-        {
-            initialState.enterState('{', HeaderListState.INSTANCE);
-        }
-
-        @Override
-        protected ModelNode fromNonDMRString(String value) throws CommandFormatException {
-            callback.reset();
-            ParserUtil.parse(value, callback, initialState);
-            final Collection<ParsedOperationRequestHeader> headers = callback.getHeaders();
-            if(headers.isEmpty()) {
-                throw new CommandFormatException("'" + value +
-                        "' doesn't follow format {[rollout server_group_list [rollback-across-groups];] (<header_name>=<header_value>;)*}");
-            }
-            final ModelNode node = new ModelNode();
-            for(ParsedOperationRequestHeader header : headers) {
-                header.addTo(null, node);
-            }
-            return node;
         }
     };
