+++ cli/src/main/java/org/jboss/as/cli/impl/CommandContextImpl.java
@@ -322,7 +322,7 @@ class CommandContextImpl implements CommandContext {
 
         // rollout plan
         final GenericTypeOperationHandler rolloutPlan = new GenericTypeOperationHandler(this, "/management-client-content=rollout-plans/rollout-plan", null);
-        rolloutPlan.addValueConverter("content", ArgumentValueConverter.HEADERS);
+        rolloutPlan.addValueConverter("content", new HeadersArgumentValueConverter(this));
         rolloutPlan.addValueCompleter("content", RolloutPlanCompleter.INSTANCE);
         cmdRegistry.registerHandler(rolloutPlan, "rollout-plan");
